etl/requirements.txt
requests==2.32.3
psycopg2-binary==2.9.9
pydantic==2.7.1
python-dotenv==1.0.1